refactor(auth): separate client component
